added score, fixed score bugs and made turrets more threatening
Added a global variable called score which keep track of the score throughout the round and displays at the end screen, the turret could kill a  ebar before, he had been mvoed to prevent the player from getting free 100 points. and the turret bullets are faster and bigger so they are more likely to hit the player when walking past them.